% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pps.R
\name{plot_ppc}
\alias{plot_ppc}
\title{Posterior predictive checks of marginals}
\usage{
plot_ppc(ppc, data, trt_name, med_name, out_name)
}
\arguments{
\item{ppc}{An objected obtained from the fitted model using the zoib_ppc function}

\item{data}{The original dataset}

\item{trt_name}{Name of the treatment variable in the original dataset}

\item{med_name}{Name of the mediator variable in the original dataset}

\item{out_name}{Name of the outcome variable in the original dataset}
}
\value{
ggplots giving the posterior probability check for the continuous
and discrete parts of the data
}
\description{
Makes plots assessing the fit of the posterior predictive distribution to the
data. Returns two figures as ggplot objects: the first gives the posterior
predictive distribution of a kernel density estimate to the continuous part
of the outcome/mediator stratified by treatment, the second gives posterior
samples of the proportions of 0s and 1s for the outcome/mediator stratified
by treatment. For both figures, the statistics computed on the original data
are provided as a baseline.
}
