% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_zoib_sensitivity.R
\name{plot_zoib_sensitivity}
\alias{plot_zoib_sensitivity}
\title{Plot sensitivity analysis for ZOIB model}
\usage{
plot_zoib_sensitivity(sensitivity)
}
\arguments{
\item{sensitivity}{Objected obtained from the zoib_sensitivity function}
}
\value{
Plot of the sensitivity analysis for different values of the sensitivity
parameter lambda, as defined by the zoib_sensitivity function.
}
\description{
Plot sensitivity analysis for ZOIB model
}
