% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_zoib.R
\name{summary_zoib}
\alias{summary_zoib}
\title{Summarize ZOIB fit}
\usage{
summary_zoib(fit, formula_y, formula_m, data)
}
\arguments{
\item{fit}{Model fitted with bayes_zoib}

\item{formula_y}{Formula used to fit the model for the outcome}

\item{formula_m}{Formula used to fit the model for the mediator}

\item{data}{The data used to fit the model}
}
\value{
A data.frame containing estimated coefficients, standard errors, etc
for the regression coefficients of the different GLMs.
}
\description{
Makes a data.frame containing estimated coefficients, standard errors, etc
for the regression coefficients of the different GLMs in the ZOIB model.
}
\details{
The "alpha" model refers to the model for the point mass at 0, the "gamma"
model is the model for the point mass at 1 (given non-zero), and "mu" and
"phi" refer to the models for the continuous part with mean mu and precision
phi.
}
