% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traceplot_zoib.R
\name{traceplot_zoib}
\alias{traceplot_zoib}
\title{Traceplots for ZOIB fit}
\usage{
traceplot_zoib(fit, formula, data, param = "alpha", response = "mediator")
}
\arguments{
\item{fit}{Model fit with bayes_zoib}

\item{formula}{Formula used to fit the model (either for the mediator or outcome, depending on the response type)}

\item{data}{Data used to fit the model}

\item{param}{The model component we want to look at; can be "alpha", "gamma",
"mu", or "phi" where alpha corresponds to the mass at 0, gamma to the mass at
 1, and (mu,phi) to the mean/precision of the beta model.}

\item{response}{"mediator" or "outcome"?}
}
\value{
A ggplot object showing traceplots.
}
\description{
Returns a traceplot for the coefficients of a fitted ZOIB model
}
