% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zoib_mediate.R
\name{zoib_mediate}
\alias{zoib_mediate}
\title{Perform causal mediation analysis in the zero-one inflated beta regression
model}
\usage{
zoib_mediate(
  zoib_fit,
  data,
  formula_m = ~1,
  formula_y = ~1,
  g_comp_thin = 1,
  print_interval = 100,
  num_copy = 2,
  med_name = "M",
  trt_name = "A",
  out_name = "Y"
)
}
\arguments{
\item{zoib_fit}{Model fit with the bayes_zoib function}

\item{data}{Original data used to get zoib_fit}

\item{formula_m}{formula for the mediator used in constructing zoib_fit}

\item{formula_y}{formula for the outcome used in constructign zoib_fit}

\item{g_comp_thin}{Thinning interval used for the g-computation, defauls to 1}

\item{print_interval}{How many iterations before printing progress?}

\item{num_copy}{Number of pseudo-datasets to use when performing g-computation}

\item{med_name}{Name of the mediator in the original data}

\item{trt_name}{Name of the treatment in the original data}

\item{out_name}{Name of the outcome in the original data}
}
\value{
A data.frame containing samples from the posterior of the direct (zeta),
indirect (delta), and total effects (tau) for different values of the treatment,
as well as the expected value of the mediator for the different treatments.

The output can be fed into zoib_sensitivity and plot_zoib_sensitivity, or
post-processed directly using the data.frame.
}
\description{
Perform causal mediation analysis in the zero-one inflated beta regression
model
}
