% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zoib_sensitivity.R
\name{zoib_sensitivity}
\alias{zoib_sensitivity}
\title{Sensitivity analysis for ZOIB mediation model}
\usage{
zoib_sensitivity(
  data,
  samples,
  formula_m = ~1,
  formula_y = ~1,
  g_comp_thin = 1,
  print_interval = 100,
  num_copy = 2,
  med_name,
  trt_name,
  out_name,
  rho = 1,
  lambda_grid,
  logit_scale = TRUE
)
}
\arguments{
\item{data}{Data used to fit the original model}

\item{samples}{Model fit obtained by bayes_zoib}

\item{formula_m}{Formula for the mediator when fitting the model}

\item{formula_y}{Formula for the outcome when fitting the model}

\item{g_comp_thin}{Do we want to sample on all iterations (thin = 1) or half (thin = 2), etc}

\item{print_interval}{Number of iterations before printing an update on progress}

\item{num_copy}{Number of pseudo-datasets when performing g-computation}

\item{med_name}{Name of the mediator}

\item{trt_name}{Name of the treatment}

\item{out_name}{Name of the outcome}

\item{rho}{Correlation in Gaussian copula (only relevant when doing sensitivity on the logit scale)}

\item{lambda_grid}{The values of lambda we want to look at}

\item{logit_scale}{Whether to operate on the logit scale or not.}
}
\value{
Samples of the mediation effects
(delta(0), delta(1), zeta(0), zeta(1), tau) for each value of lambda, stored
as a data.frame with the iteration, id of the simulated dataset, parameter,
and value of lambda given.

This can be worked with directly for inference, but it is probably more efficient to
visualize the results with plot_zoib_sensitivity.
}
\description{
Performs the sensitivity analysis outlined by Rene et al. (2021) Causal
Mediation and Sensitivity Analysis for Mixed-Scale Data - see this paper for
additional details. This sensitivity analysis shifts unobserved potential
outcomes Y(a, m) either on the linear or logit scale by a factor lambda *
(M(a) - m), where lambda is an unidentified sensitivity parameter.
}
